+++ src/assault/mod.rs
@@ -6,6 +6,6 @@ pub mod messages;
 pub mod service;
 
 pub mod measure;
-pub mod metrics;
+pub mod error;
 pub mod reportable;
 pub mod worker;
